requirements.txt
numpy==1.23.3
google-cloud-vision==3.1.3
Flask==2.2.2
googletrans==4.0.0rc1